feat: recover client wrapper cmd
